requirements.txt
alembic
black
click
cryptography
deepdiff
flake8
flask
flask-cors
flask-smorest
git+https://github.com/Johnson-Gage-Inspection-Inc/qualer-sdk-python.git@13a87a64dccfe9cfeab18e57d6b59b7ed0c79335#egg=qualer_sdk
isort
marshmallow
msal
Office365-REST-Python-Client==2.6.2
openpyxl
pandas
pre-commit
psycopg2-binary
pydantic
pyjwt
pytest
pytest-cov
pytest-mock
python-dotenv
requests
sqlalchemy
xlrd